Use standard std::exp/std::log float overloads for portability
std::expf and std::logf are not reliably declared by libstdc++; the
float overloads of std::exp/std::log resolve identically for float
arguments and compile everywhere.

+++ src/transformer/tensor.cpp
@@ -672,7 +672,7 @@ Tensor Tensor::softmax() const {
             float sum = 0.0f;
             #pragma clang loop vectorize(enable)
             for (size_t j = 0; j < this->cols; j++) {
-                const float val = std::expf(row_in[j] - max_val);
+                const float val = std::exp(row_in[j] - max_val);
                 row_out[j] = val;
                 sum += val;
             }
@@ -705,7 +705,7 @@ Tensor Tensor::softmax() const {
                 float sum = 0.0f;
                 #pragma clang loop vectorize(enable)
                 for (size_t j = 0; j < this->cols; j++) {
-                    const float val = std::expf(row_in[j] - max_val);
+                    const float val = std::exp(row_in[j] - max_val);
                     row_out[j] = val;
                     sum += val;
                 }
